BR-Task1-Refactoring
